Move component to analytics
