+++ apps/hyperdrive-trading/src/ui/analytics/ExternalLink.tsx
