+++ apps/hyperdrive-trading/src/ui/app/Navbar/Navbar.tsx
@@ -7,9 +7,9 @@ import { Link, useRouterState } from "@tanstack/react-router";
 import classNames from "classnames";
 import { ReactElement } from "react";
 import { isTestnetChain } from "src/chains/isTestnetChain";
+import { ExternalLink } from "src/ui/analytics/ExternalLink";
 import { DevtoolsMenu } from "src/ui/app/Navbar/DevtoolsMenu";
 import { HyperdriveLogo } from "src/ui/app/Navbar/HyperdriveLogo";
-import { ExternalLink } from "src/ui/base/components/ExternalLink";
 import VersionPicker from "src/ui/base/components/VersionPicker";
 import { useIsTailwindSmallScreen } from "src/ui/base/mediaBreakpoints";
 import { useRegionInfo } from "src/ui/compliance/hooks/useRegionInfo";
